requirements.txt
fastapi
uvicorn
jinja2
python-multipart
sqlalchemy
aiofiles
pillow
gunicorn
locust
PyMySQL
cryptography
opencv-python

